use int64 texture ids everywhere

+++ src/osc/Bundler/ChunkedImageBundler.cpp
@@ -2,7 +2,7 @@
 #include "../ChunkedSend/ChunkedImage.hpp"
 
 ChunkedImageBundler::ChunkedImageBundler(
-  int32_t chunkedSendId,
+  int64_t chunkedSendId,
   int32_t chunkNum,
   int32_t numChunks,
   int32_t chunkSize,
